+++ R/stat_compare_means.R
@@ -7,6 +7,9 @@ NULL
 #'@inheritParams compare_means
 #'@param method a character string indicating which method to be used for
 #'  comparing means.
+#'@param method.args a list of additional arguments used for the test method.
+#'  For example one might use \code{method.args = list(alternative = "greater")}
+#'  for wilcoxon test.
 #'@param comparisons A list of length-2 vectors. The entries in the vector are
 #'  either the names of 2 values on the x-axis or the 2 integers that correspond
 #'  to the index of the groups of interest, to be compared.
@@ -92,7 +95,7 @@ NULL
 #'
 #'@export
 stat_compare_means <- function(mapping = NULL, data = NULL,
-                     method = NULL, paired = FALSE, ref.group = NULL,
+                     method = NULL, paired = FALSE, method.args = list(), ref.group = NULL,
                      comparisons = NULL, hide.ns = FALSE, label.sep = ", ",
                      label = NULL, label.x.npc = "left", label.y.npc = "top",
                      label.x = NULL, label.y = NULL, tip.length = 0.03,
@@ -105,9 +108,10 @@ stat_compare_means <- function(mapping = NULL, data = NULL,
     method.info <- .method_info(method)
     method <- method.info$method
 
-    test.args <- list(paired = paired)
+    method.args <- .add_item(method.args, paired = paired)
     if(method == "wilcox.test")
-      test.args$exact <- FALSE
+      method.args$exact <- FALSE
+
 
     pms <- list(...)
     size <- ifelse(is.null(pms$size), 0.3, pms$size)
@@ -124,7 +128,7 @@ stat_compare_means <- function(mapping = NULL, data = NULL,
 
     step_increase <- ifelse(is.null(label.y), 0.12, 0)
     ggsignif::geom_signif(comparisons = comparisons, y_position = label.y,
-                          test = method, test.args = test.args,
+                          test = method, test.args = method.args,
                           step_increase = step_increase, size = size, color = color,
                           map_signif_level = map_signif_level, tip_length = tip.length)
   }
@@ -136,7 +140,8 @@ stat_compare_means <- function(mapping = NULL, data = NULL,
       position = position, show.legend = show.legend, inherit.aes = inherit.aes,
       params = list(label.x.npc  = label.x.npc , label.y.npc  = label.y.npc,
                     label.x = label.x, label.y = label.y, label.sep = label.sep,
-                    method = method, paired = paired, ref.group = ref.group,
+                    method = method, method.args = method.args,
+                    paired = paired, ref.group = ref.group,
                     symnum.args = symnum.args,
                     hide.ns = hide.ns, na.rm = na.rm, ...)
     )
@@ -150,8 +155,8 @@ StatCompareMeans<- ggproto("StatCompareMeans", Stat,
                   required_aes = c("x", "y"),
                   default_aes = aes(hjust = ..hjust.., vjust = ..vjust..),
 
-                  compute_panel = function(data, scales, method, paired, ref.group,
-                                           symnum.args,
+                  compute_panel = function(data, scales, method, method.args,
+                                           paired, ref.group, symnum.args,
                                            hide.ns, label.x.npc, label.y.npc,
                                            label.x, label.y, label.sep)
                     {
@@ -179,16 +184,20 @@ StatCompareMeans<- ggproto("StatCompareMeans", Stat,
 
                     # Perform group comparisons
                     #::::::::::::::::::::::::::::::::::::::::::::::::::
-                    if(.is.multiple.grouping.vars){
-                      .test <- compare_means(y~group, data = data, method = method, group.by = "x",
-                                             paired = paired, ref.group = ref.group,
-                                             symnum.args = symnum.args)
+                    method.args <- method.args %>%
+                      .add_item(data = data, method = method,
+                                paired = paired, ref.group = ref.group,
+                                symnum.args = symnum.args)
 
+                    if(.is.multiple.grouping.vars){
+                      method.args <- method.args %>%
+                        .add_item(formula = y ~ group, group.by = "x")
+                      .test <- do.call(compare_means, method.args)
                     }
                     else{
-                      .test <- compare_means(y~x, data = data, method = method,
-                                             paired = paired, ref.group = ref.group,
-                                             symnum.args = symnum.args)
+                      method.args <- method.args %>%
+                        .add_item(formula = y ~ x)
+                      .test <- do.call(compare_means, method.args)
                     }
 
                     pvaltxt <- ifelse(.test$p < 2.2e-16, "p < 2.2e-16",